refactor: modularize data enhancer
