+++ components/plugin_adapter.py
@@ -8,7 +8,7 @@ from typing import Any, Dict, List, Optional
 import pandas as pd
 
 from core.plugins.service_locator import PluginServiceLocator
-from services.data_enhancer import get_ai_column_suggestions
+from services.data_enhancer.mapping_utils import get_ai_column_suggestions
 
 logger = logging.getLogger(__name__)
