+++ services/ai_suggestions.py
@@ -1,7 +1,7 @@
 import logging
 from typing import Any, Dict, Sequence
 
-from services.data_enhancer import get_ai_column_suggestions
+from services.data_enhancer.mapping_utils import get_ai_column_suggestions
 
 logger = logging.getLogger(__name__)
