+++ services/data_enhancer/__init__.py
@@ -1,0 +1,17 @@
+"""Data enhancement utilities and CLI entrypoint."""
+
+from .mapping_utils import (
+    apply_fuzzy_column_matching,
+    apply_manual_mapping,
+    get_ai_column_suggestions,
+    get_mapping_suggestions,
+)
+from .cli import run_data_enhancer
+
+__all__ = [
+    "apply_fuzzy_column_matching",
+    "apply_manual_mapping",
+    "get_ai_column_suggestions",
+    "get_mapping_suggestions",
+    "run_data_enhancer",
+]
