+++ services/upload/core/orchestrator.py
@@ -6,7 +6,7 @@ from typing import Any, Callable, Dict, List, Tuple
 import pandas as pd
 
 from services.async_file_processor import AsyncFileProcessor
-from services.data_enhancer import get_ai_column_suggestions
+from services.data_enhancer.mapping_utils import get_ai_column_suggestions
 from services.upload.core.file_processor_service import FileProcessor
 from services.upload.core.learning_coordinator import LearningCoordinator
 from services.upload.core.ui_builder import UploadUIBuilder
